[OpenMP][NFC] Merge elf_common into PluginInterface
The overhead of a library and 4 files seems high without benefit.
This simply tries to consolidate our structure.
