+++ openmp/libomptarget/plugins-nextgen/common/CMakeLists.txt
@@ -13,4 +13,3 @@
 add_subdirectory(OMPT)
 add_subdirectory(PluginInterface)
 add_subdirectory(MemoryManager)
-add_subdirectory(elf_common)
